Enum gtk_sys::GtkInputPurpose

pub enum GtkInputPurpose {
    FreeForm,
    Alpha,
    Digits,
    Number,
    Phone,
    Url,
    Email,
    Name,
    Password,
    Pin,
}

Describes primary purpose of the input widget. This information is useful for on-screen keyboards and similar input methods to decide which keys should be presented to the user.

Note that the purpose is not meant to impose a totally strict rule about allowed characters, and does not replace input validation. It is fine for an on-screen keyboard to let the user override the character set restriction that is expressed by the purpose. The application is expected to validate the entry contents, even if it specified a purpose.

The difference between InputPurpose::Digits and InputPurpose::Number is that the former accepts only digits while the latter also some punctuation (like commas or points, plus, minus) and “e” or “E” as in 3.14E+000.

This enumeration may be extended in the future; input methods should interpret unknown values as “free form”.

Variants

FreeForm

Allow any character

Alpha

Allow only alphabetic characters

Digits

Allow only digits

Number

Edited field expects numbers

Phone

Edited field expects phone number

Url

Edited field expects URL

Email

Edited field expects email address

Name

Edited field expects the name of a person

Password

Like InputPurpose::FreeForm, but characters are hidden

Pin

Like InputPurpose::Digits, but characters are hidden

Since: 3.6